Try: Distinguish "backwards compatible" CSS rules
Here, a "backwards compatible rule" is a rule made up of two selectors
that only differ in the fact that one targets .editor- (deprecated) and
the other .block-editor- (current). For example:

.editor-block-list__foo bar,
.block-editor-block-list__foo bar {
  /* ... */
}

This distinction can be used to communicate CSS violations to the user
using different urgency levels (console.warn, console.info).